data layer implemented

+++ JobFinder/src/main/java/com/ofdun/jobfinder/features/clients/ai/AiClient.java
@@ -3,5 +3,5 @@ package com.ofdun.jobfinder.features.clients.ai;
 import java.util.List;
 
 public interface AiClient {
-    List<Double> getEmbedding(String content);
+    List<Float> getEmbedding(String content);
 }
