+++ JobFinder/src/main/java/com/ofdun/jobfinder/features/matching/service/BasicMatchingService.java
@@ -1,15 +1,14 @@
 package com.ofdun.jobfinder.features.matching.service;
 
 import com.ofdun.jobfinder.features.clients.ai.AiClient;
-import com.ofdun.jobfinder.shared.matching.domain.model.MatchResultModel;
 import com.ofdun.jobfinder.features.resume.domain.repository.VectorResumeRepository;
 import com.ofdun.jobfinder.features.vacancy.domain.repository.VacancyRepository;
+import com.ofdun.jobfinder.shared.matching.model.MatchResultModel;
+import java.util.Collections;
+import java.util.List;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.Collections;
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -19,7 +18,8 @@ public class BasicMatchingService implements MatchingService {
     private final AiClient aiClient;
 
     @Override
-    public List<MatchResultModel> findSuitableCandidates(@NonNull Long vacancyId, @NonNull Integer maxAmount) {
+    public List<MatchResultModel> findSuitableCandidates(
+            @NonNull Long vacancyId, @NonNull Integer maxAmount) {
         if (maxAmount <= 0) {
             return Collections.emptyList();
         }
@@ -28,7 +28,7 @@ public class BasicMatchingService implements MatchingService {
         return vectorResumeRepository.getMostSimilarResumes(vacancyEmbedding, maxAmount);
     }
 
-    private List<Double> getVacancyEmbedding(@NonNull Long vacancyId) {
+    private List<Float> getVacancyEmbedding(@NonNull Long vacancyId) {
         var vacancy = vacancyRepository.getVacancyById(vacancyId);
         return aiClient.getEmbedding(vacancy.toString());
     }
