+++ JobFinder/src/main/java/com/ofdun/jobfinder/features/resume/domain/repository/VectorResumeRepository.java
@@ -1,14 +1,17 @@
 package com.ofdun.jobfinder.features.resume.domain.repository;
 
 import com.ofdun.jobfinder.features.resume.domain.model.ResumeModel;
-import com.ofdun.jobfinder.shared.matching.domain.model.MatchResultModel;
-
+import com.ofdun.jobfinder.shared.matching.model.MatchResultModel;
 import java.util.List;
 
 public interface VectorResumeRepository {
     Long createResume(ResumeModel resumeModel);
+
     ResumeModel getResumeById(Long resumeId);
+
     ResumeModel updateResume(ResumeModel resumeModel);
-    List<MatchResultModel> getMostSimilarResumes(List<Double> embedding, Integer maxAmount);
+
+    List<MatchResultModel> getMostSimilarResumes(List<Float> embedding, Integer maxAmount);
+
     Boolean deleteResume(Long resumeId);
 }
